requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
colorama==0.4.6
    # via tqdm
idna==3.4
    # via requests
numpy==1.21.6
    # via
    #   -r requirements.in
    #   opencv-python
    #   pandas
    #   torchvision
opencv-python==4.6.0.66
    # via -r requirements.in
pandas==1.3.5
    # via -r requirements.in
pillow==9.3.0
    # via
    #   -r requirements.in
    #   torchvision
python-dateutil==2.8.2
    # via pandas
pytz==2022.6
    # via pandas
requests==2.28.1
    # via torchvision
six==1.16.0
    # via python-dateutil
torch==1.13.0
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.14.0
    # via -r requirements.in
tqdm==4.64.1
    # via -r requirements.in
typing-extensions==4.4.0
    # via
    #   torch
    #   torchvision
urllib3==1.26.12
    # via requests
